refactor: Rename dto and vo of CreateBoard
