+++ src/board/dto/CreateBoard.dto.ts
