+++ src/board/vo/CreateBoard.vo.ts
@@ -3,7 +3,7 @@ import {BoardStatus} from "../BoardStatus.model";
 export class CreateBoardVo {
     readonly title: string;
     readonly description: string;
-    // 값 강제가 왜 안되지?
+    // 값 강제가 안됨.
     readonly status: keyof typeof BoardStatus;
 
     constructor(title: string, description: string, status: keyof typeof BoardStatus) {
